backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prisma==0.15.0
pydantic==2.4.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
